refactor spleeter, must test
