+++ spleeter/requirements.txt
@@ -1,1 +1,3 @@
 pika==1.1.0
+pydub==0.24.1
+Spleeter==2.0.1
